Improve test on xlsx
